Move hekad config into pipeline config.
Watch for signals in separate go routine.
