+++ cmd/hekad/main.go
@@ -40,50 +40,13 @@ import (
 	_ "github.com/mozilla-services/heka/plugins/statsd"
 	_ "github.com/mozilla-services/heka/plugins/tcp"
 	_ "github.com/mozilla-services/heka/plugins/udp"
-	"io/ioutil"
-	"log"
 	"os"
 	"path/filepath"
-	"runtime"
-	"runtime/pprof"
-	"strconv"
-	"strings"
-	"syscall"
 )
 
 const (
 	VERSION = "0.8.0"
 )
-
-func setGlobalConfigs(config *HekadConfig) (*pipeline.GlobalConfigStruct, string, string) {
-	maxprocs := config.Maxprocs
-	poolSize := config.PoolSize
-	chanSize := config.ChanSize
-	cpuProfName := config.CpuProfName
-	memProfName := config.MemProfName
-	maxMsgLoops := config.MaxMsgLoops
-	maxMsgProcessInject := config.MaxMsgProcessInject
-	maxMsgProcessDuration := config.MaxMsgProcessDuration
-	maxMsgTimerInject := config.MaxMsgTimerInject
-
-	runtime.GOMAXPROCS(maxprocs)
-
-	globals := pipeline.DefaultGlobals()
-	globals.PoolSize = poolSize
-	globals.PluginChanSize = chanSize
-	globals.MaxMsgLoops = maxMsgLoops
-	if globals.MaxMsgLoops == 0 {
-		globals.MaxMsgLoops = 1
-	}
-	globals.MaxMsgProcessInject = maxMsgProcessInject
-	globals.MaxMsgProcessDuration = maxMsgProcessDuration
-	globals.MaxMsgTimerInject = maxMsgTimerInject
-	globals.BaseDir = config.BaseDir
-	globals.ShareDir = config.ShareDir
-	globals.SampleDenominator = config.SampleDenominator
-
-	return globals, cpuProfName, memProfName
-}
 
 func main() {
 	configPath := flag.String("config", filepath.FromSlash("/etc/hekad.toml"),
@@ -91,11 +54,6 @@ func main() {
 			"in the directory will be loaded.")
 	version := flag.Bool("version", false, "Output version and exit")
 	flag.Parse()
-
-	config := &HekadConfig{}
-	var err error
-	var cpuProfName string
-	var memProfName string
 
 	if flag.NFlag() == 0 {
 		flag.PrintDefaults()
@@ -107,112 +65,5 @@ func main() {
 		os.Exit(0)
 	}
 
-	config, err = LoadHekadConfig(*configPath)
-	if err != nil {
-		log.Fatal("Error reading config: ", err)
-	}
-	if config.SampleDenominator <= 0 {
-		log.Fatalln("'sample_denominator' value must be greater than 0.")
-	}
-	globals, cpuProfName, memProfName := setGlobalConfigs(config)
-
-	if err = os.MkdirAll(globals.BaseDir, 0755); err != nil {
-		log.Fatalf("Error creating 'base_dir' %s: %s", config.BaseDir, err)
-	}
-
-	if config.PidFile != "" {
-		contents, err := ioutil.ReadFile(config.PidFile)
-		if err == nil {
-			pid, err := strconv.Atoi(strings.TrimSpace(string(contents)))
-			if err != nil {
-				log.Fatalf("Error reading proccess id from pidfile '%s': %s", config.PidFile, err)
-			}
-
-			process, err := os.FindProcess(pid)
-
-			// on Windows, err != nil if the process cannot be found
-			if runtime.GOOS == "windows" {
-				if err == nil {
-					log.Fatalf("Process %d is already running.", pid)
-				}
-			} else if process != nil {
-				// err is always nil on POSIX, so we have to send the process
-				// a signal to check whether it exists
-				if err = process.Signal(syscall.Signal(0)); err == nil {
-					log.Fatalf("Process %d is already running.", pid)
-				}
-			}
-		}
-		if err = ioutil.WriteFile(config.PidFile, []byte(strconv.Itoa(os.Getpid())),
-			0644); err != nil {
-
-			log.Fatalf("Unable to write pidfile '%s': %s", config.PidFile, err)
-		}
-		log.Printf("Wrote pid to pidfile '%s'", config.PidFile)
-		defer func() {
-			if err = os.Remove(config.PidFile); err != nil {
-				log.Printf("Unable to remove pidfile '%s': %s", config.PidFile, err)
-			}
-		}()
-	}
-
-	if cpuProfName != "" {
-		profFile, err := os.Create(cpuProfName)
-		if err != nil {
-			log.Fatalln(err)
-		}
-
-		pprof.StartCPUProfile(profFile)
-		defer func() {
-			pprof.StopCPUProfile()
-			profFile.Close()
-		}()
-	}
-
-	if memProfName != "" {
-		defer func() {
-			profFile, err := os.Create(memProfName)
-			if err != nil {
-				log.Fatalln(err)
-			}
-			pprof.WriteHeapProfile(profFile)
-			profFile.Close()
-		}()
-	}
-
-	// Set up and load the pipeline configuration and start the daemon.
-	pipeconf := pipeline.NewPipelineConfig(globals)
-	if err = loadFullConfig(pipeconf, configPath); err != nil {
-		log.Fatal("Error reading config: ", err)
-	}
-	pipeline.Run(pipeconf)
-}
-
-func loadFullConfig(pipeconf *pipeline.PipelineConfig, configPath *string) (err error) {
-	p, err := os.Open(*configPath)
-	if err != nil {
-		return fmt.Errorf("error opening file: %s", err.Error())
-	}
-	fi, err := p.Stat()
-	if err != nil {
-		return fmt.Errorf("can't stat file: %s", err.Error())
-	}
-
-	if fi.IsDir() {
-		files, _ := ioutil.ReadDir(*configPath)
-		for _, f := range files {
-			fName := f.Name()
-			if !strings.HasSuffix(fName, ".toml") {
-				// Skip non *.toml files in a config dir.
-				continue
-			}
-			err = pipeconf.LoadFromConfigFile(filepath.Join(*configPath, fName))
-			if err != nil {
-				break
-			}
-		}
-	} else {
-		err = pipeconf.LoadFromConfigFile(*configPath)
-	}
-	return
+	pipeline.Start(*configPath)
 }
