+++ pipeline/pipeline_runner.go
@@ -19,10 +19,14 @@ package pipeline
 import (
 	"github.com/mozilla-services/heka/message"
 	"github.com/rafrombrc/go-notify"
+	"io/ioutil"
 	"log"
 	"os"
 	"os/signal"
 	"path/filepath"
+	"runtime"
+	"runtime/pprof"
+	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -191,6 +195,117 @@ func (p *PipelinePack) Recycle() {
 	}
 }
 
+func WatchSignals(config *PipelineConfig, signalChan chan os.Signal) {
+	globals := config.Globals
+	for !globals.IsShuttingDown() {
+		select {
+		case sig := <-globals.sigChan:
+			switch sig {
+			case syscall.SIGHUP:
+				log.Println("Reload initiated.")
+				if err := notify.Post(RELOAD, nil); err != nil {
+					log.Println("Error sending reload event: ", err)
+				}
+				signalChan <- sig
+			case syscall.SIGINT, syscall.SIGTERM:
+				log.Println("Shutdown initiated.")
+				globals.stop()
+				signalChan <- sig
+			case SIGUSR1:
+				log.Println("Queue report initiated.")
+				go config.allReportsStdout()
+			}
+
+		}
+	}
+}
+
+func Start(configPath string) {
+	config := &HekadConfig{}
+	var err error
+	var cpuProfName string
+	var memProfName string
+
+	config, err = LoadHekadConfig(configPath)
+	if err != nil {
+		log.Fatal("Error reading config: ", err)
+	}
+	if config.SampleDenominator <= 0 {
+		log.Fatalln("'sample_denominator' value must be greater than 0.")
+	}
+	globals, cpuProfName, memProfName := setGlobalConfigs(config)
+
+	if err = os.MkdirAll(globals.BaseDir, 0755); err != nil {
+		log.Fatalf("Error creating 'base_dir' %s: %s", config.BaseDir, err)
+	}
+
+	if config.PidFile != "" {
+		contents, err := ioutil.ReadFile(config.PidFile)
+		if err == nil {
+			pid, err := strconv.Atoi(strings.TrimSpace(string(contents)))
+			if err != nil {
+				log.Fatalf("Error reading proccess id from pidfile '%s': %s", config.PidFile, err)
+			}
+
+			process, err := os.FindProcess(pid)
+
+			// on Windows, err != nil if the process cannot be found
+			if runtime.GOOS == "windows" {
+				if err == nil {
+					log.Fatalf("Process %d is already running.", pid)
+				}
+			} else if process != nil {
+				// err is always nil on POSIX, so we have to send the process
+				// a signal to check whether it exists
+				if err = process.Signal(syscall.Signal(0)); err == nil {
+					log.Fatalf("Process %d is already running.", pid)
+				}
+			}
+		}
+		if err = ioutil.WriteFile(config.PidFile, []byte(strconv.Itoa(os.Getpid())),
+			0644); err != nil {
+
+			log.Fatalf("Unable to write pidfile '%s': %s", config.PidFile, err)
+		}
+		log.Printf("Wrote pid to pidfile '%s'", config.PidFile)
+		defer func() {
+			if err = os.Remove(config.PidFile); err != nil {
+				log.Printf("Unable to remove pidfile '%s': %s", config.PidFile, err)
+			}
+		}()
+	}
+
+	if cpuProfName != "" {
+		profFile, err := os.Create(cpuProfName)
+		if err != nil {
+			log.Fatalln(err)
+		}
+
+		pprof.StartCPUProfile(profFile)
+		defer func() {
+			pprof.StopCPUProfile()
+			profFile.Close()
+		}()
+	}
+
+	if memProfName != "" {
+		defer func() {
+			profFile, err := os.Create(memProfName)
+			if err != nil {
+				log.Fatalln(err)
+			}
+			pprof.WriteHeapProfile(profFile)
+			profFile.Close()
+		}()
+	}
+	// Set up and load the pipeline configuration and start the daemon.
+	pConfig := NewPipelineConfig(globals)
+	if err = LoadFullConfig(pConfig, configPath); err != nil {
+		log.Fatal("Error reading config: ", err)
+	}
+	Run(pConfig)
+}
+
 // Main function driving Heka execution. Loads config, initializes
 // PipelinePack pools, and starts all the runners. Then it listens for signals
 // and drives the shutdown process when that is triggered.
@@ -201,6 +316,11 @@ func Run(config *PipelineConfig) {
 	var err error
 
 	globals := config.Globals
+	signal.Notify(globals.sigChan, syscall.SIGINT, syscall.SIGTERM, syscall.SIGHUP, SIGUSR1)
+
+	signalChan := make(chan os.Signal)
+	defer close(signalChan)
+	go WatchSignals(config, signalChan)
 
 	for name, output := range config.OutputRunners {
 		outputsWg.Add(1)
@@ -254,27 +374,8 @@ func Run(config *PipelineConfig) {
 		log.Printf("Input started: %s\n", name)
 	}
 
-	// wait for sigint
-	signal.Notify(globals.sigChan, syscall.SIGINT, syscall.SIGTERM, syscall.SIGHUP, SIGUSR1)
-
-	for !globals.IsShuttingDown() {
-		select {
-		case sig := <-globals.sigChan:
-			switch sig {
-			case syscall.SIGHUP:
-				log.Println("Reload initiated.")
-				if err := notify.Post(RELOAD, nil); err != nil {
-					log.Println("Error sending reload event: ", err)
-				}
-			case syscall.SIGINT, syscall.SIGTERM:
-				log.Println("Shutdown initiated.")
-				globals.stop()
-			case SIGUSR1:
-				log.Println("Queue report initiated.")
-				go config.allReportsStdout()
-			}
-		}
-	}
+	// wait for a signal
+	<-signalChan
 
 	config.inputsLock.Lock()
 	for _, input := range config.InputRunners {
